requirements.txt
asgiref==3.8.1
Django==5.1.6
django-filter==25.1
djangorestframework==3.15.2
Markdown==3.7
psycopg==3.2.4
sqlparse==0.5.3
typing_extensions==4.12.2
